% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containerOpt.R
\name{containerOpt}
\alias{containerOpt}
\title{Create container options}
\usage{
containerOpt(Image, Hostname = "", Domainname = "", User = "",
  Memory = 0, MemorySwap = 0, CpuShares = 0, Cpuset = "",
  AttachStdin = FALSE, AttachStdout = TRUE, AttachStderr = TRUE,
  PortSpecs = NA,
  ExposedPorts = HostConfig[["PortBindings"]][["container"]], Tty = FALSE,
  OpenStdin = FALSE, StdinOnce = FALSE, Env = list(), Cmd = NA,
  Volumes = NULL, WorkingDir = "", Entrypoint = NA,
  NetworkDisabled = FALSE, MacAddress = "", OnBuild = NA,
  HostConfig = hostConfig())
}
\arguments{
\item{Image}{String value containing the image name to use for the container. Corresponds to docker run IMAGE.}

\item{Hostname}{A string value containing the desired hostname to use for the container. Corresponds to docker run -h hostname.domain.com.}

\item{Domainname}{A string value containing the desired domainname to use for the container. Corresponds to docker run -h hostname.domain.com.
\describe{
\item{Example:}{As as example: \code{Hostname = "hostname", Domainname = "domain.com"} would correspond to \code{docker run -h hostname.domain.com}}
}}

\item{User}{A string value containg the user to use inside the container. Corresponds to docker run -u}

\item{Memory}{Memory limit in bytes. Corresponds to docker run -m}

\item{MemorySwap}{Total memory usage (memory + swap); set -1 to disable swap.}

\item{CpuShares}{An integer value containing the CPU Shares for container (ie. the relative weight vs othercontainers). Corresponds to docker run -c}

\item{Cpuset}{String value containg the cgroups Cpuset to use. Corresponds to docker run --cpuset=""}

\item{AttachStdin}{Boolean value, attaches to stdin. Corresponds to docker run -a stdin}

\item{AttachStdout}{Boolean value, attaches to stdin. Corresponds to docker run -a stdout}

\item{AttachStderr}{Boolean value, attaches to stdin. Corresponds to docker run -a stderr}

\item{PortSpecs}{ADD DOCU}

\item{ExposedPorts}{An object mapping ports to an empty object in the form of: "ExposedPorts": { "<port>/<tcp|udp>: {}" }. See \url{http://stackoverflow.com/questions/20428302/binding-a-port-to-a-host-interface-using-the-rest-api}}

\item{Tty}{Boolean value, Attach standard streams to a tty, including stdin if it is not closed. Corresponds to docker run -t}

\item{OpenStdin}{Boolean value, opens stdin. Corresponds to docker run -i.}

\item{StdinOnce}{Boolean value, opens stdin. Corresponds to docker run -i.}

\item{Env}{A list of environment variables in the form of VAR=value. Corresponds to docker run -e
\describe{
\item{Example:}{As an example: \code{Env = list(MYVAR1=foo1, MYVAR2=foo2)} would correspond to \code{docker run -e MYVAR1=foo1 -e MYVAR2=foo2}}
}}

\item{Cmd}{Command to run specified as a string or an array of strings.
\describe{
 \item{Example:}{As an example \code{Image = "ubuntu", Cmd = list("/bin/echo","hello","world")} would correspond to "\code{docker run ubuntu /bin/echo hello world}}
 }}

\item{Volumes}{An object mapping mountpoint paths (strings) inside the container to empty objects.}

\item{WorkingDir}{A string value containing the working dir for commands to run in. Corresponds to docker run -w}

\item{Entrypoint}{Set the entrypoint for the container as a string or an array of strings. Corresponds to docker run --entrypoint=""}

\item{NetworkDisabled}{Boolean value, when true disables neworking for the container. Corresponds to}

\item{MacAddress}{Container MAC address. Corresponds to docker run --mac-address=...}

\item{OnBuild}{ADD DOCU}

\item{HostConfig}{The host configuration. See \code{\link{hostConfig}}}
}
\description{
\code{containerOpt}
A utility function to create options suitable for creating a container.
}
\examples{
\dontrun{
containerOpt()
}
}

