% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{dockerAPI-class}
\alias{dockerAPI}
\alias{dockerAPI-class}
\title{CLASS dockerAPI}
\description{
dockerAPI Class uses the
}
\details{
dockerAPI is a generator object.
}
\section{Methods}{

\describe{
\item{\code{containerLogs(id, follow = FALSE, stdout = FALSE, stderr = FALSE,
  timestamps = FALSE, tail = "all")}}{Get stdout and stderr logs from the container id
\describe{
\item{\code{id}:}{Container id.}
\item{\code{follow}:}{1/True/true or 0/False/false, return stream. Default false}
\item{\code{stdout}:}{1/True/true or 0/False/false, show stdout log. Default false}
\item{\code{stderr}:}{1/True/true or 0/False/false, show stderr log. Default false}
\item{\code{timestamps}:}{1/True/true or 0/False/false, print timestamps for every log line. Default false}
\item{\code{tail}:}{Output specified number of lines at the end of logs: all or <number>. Default all}
}}

\item{\code{containerProcesses(id, ps_args = NULL)}}{List processes running inside the container id
\describe{
\item{\code{id}:}{Container id.}
\item{\code{ps_args}:}{ps arguments to use (e.g., aux). docker os dependent see \url{https://github.com/docker/docker/issues/8075}}
}}

\item{\code{get(all = TRUE, limit = NULL, since = NULL, before = NULL,
  size = NULL)}}{List containers:
\describe{
\item{\code{all}:}{1/True/true or 0/False/false, Show all containers. Only running containers are shown by default (i.e., this defaults to false)}
\item{\code{limit}:}{Show limit last created containers, include non-running ones.}
\item{\code{since}:}{Show only containers created since Id, include non-running ones.}
\item{\code{before}:}{Show only containers created before Id, include non-running ones.}
\item{\code{size}:}{1/True/true or 0/False/false, Show the containers sizes}
}}

\item{\code{inspectContainer(id)}}{Return low-level information on the container id
\describe{
\item{\code{id}:}{Container id.}
}}
}}
\examples{
\dontrun{
docker <- dockerAPI(ip = "192.168.59.103", port = 2375L) # windows example
docker$get()
}
}

