% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{dockerContainer-class}
\alias{dockerContainer}
\alias{dockerContainer-class}
\title{docker Class uses the}
\description{
dockerContainer is a generator object.
}
\section{Methods}{

\describe{
\item{\code{inspect()}}{Return low-level information on the container}

\item{\code{listProcesses(ps_args = NULL)}}{List processes running inside the container
\describe{
\item{\code{ps_args}:}{ps arguments to use (e.g., aux). docker os dependent see \url{https://github.com/docker/docker/issues/8075}}
}}
}}
\examples{
\dontrun{
}
}

