% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/errorHandler.R
\docType{class}
\name{errorHandler-class}
\alias{errorHandler}
\alias{errorHandler-class}
\title{CLASS errorHandler}
\description{
errorHandler Class uses the
}
\details{
errorHandler is a generator object.
}
\examples{
\dontrun{
}
}

