% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containerOpt.R
\name{hostConfig}
\alias{hostConfig}
\title{Create host configuration options}
\usage{
hostConfig(Binds = NULL, ContainerIDFile = "", LxcConf = list(),
  Privileged = FALSE, PortBindings = list(), Links = NULL,
  PublishAllPorts = FALSE, Dns = NULL, DnsSearch = NULL,
  ExtraHosts = NULL, VolumesFrom = NULL, Devices = list(),
  NetworkMode = "Bridge", IpcMode = "", CapAdd = NULL, CapDrop = NULL,
  RestartPolicy = list(Name = "", MaximumRetryCount = 0L),
  SecurityOpt = NULL)
}
\description{
\code{hostConfig}
A utility function to create host configuration options suitable for the HostConfig argument when creating a container.
}
\examples{
\dontrun{
hostConfig()
}
}

