% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containerOpt.R
\name{portBindings}
\alias{portBindings}
\title{Create Port Bindings options}
\usage{
portBindings(ipHostPort = NULL, containerPort = NULL)
}
\arguments{
\item{ipHostPort}{host A character vector of host interface and host port listings}

\item{containerPort}{A character vector of container ports. Format (80, 80/tcp 80/udp etc)}
}
\description{
\code{portBindings}
A utility function to create Port Bindings options suitable for the PortBindings argument in \code{\link{hostConfig()}}.
}
\examples{
\dontrun{
portBindings()
}
}

