% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containerOpt.R
\name{containerOpt}
\alias{containerOpt}
\title{Create container options}
\usage{
containerOpt()
}
\description{
\code{containerOpt}
A utility function to create options suitable for creating a container.
}
\examples{
\dontrun{
containerOpt()
}
}

