% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/containerOpt.R
\name{containerOpt}
\alias{containerOpt}
\title{Create container options}
\usage{
containerOpt(Image, Hostname = "", Domainname = "", User = "",
  Memory = 0, MemorySwap = 0, CpuShares = 0, Cpuset = "",
  AttachStdin = FALSE, AttachStdout = TRUE, AttahcStderr = TRUE,
  PortSpecs = NULL, ExposedPorts = list(), Tty = FALSE,
  OpenStdin = FALSE, StdinOnce = FALSE, Env = list(), Cmd = NULL,
  Volumes = list(), WorkingDir = "", Entrypoint = NULL,
  NetworkDisabled = FALSE, MacAddress = "", OnBuild = NULL,
  HostConfig = hostconfig())
}
\arguments{
\item{Image}{String value containing the image name to use for the container. Corresponds to docker run IMAGE.}

\item{Hostname}{A string value containing the desired hostname to use for the container. Corresponds to docker run -h hostname.domain.com.}

\item{Domainname}{A string value containing the desired domainname to use for the container. Corresponds to docker run -h hostname.domain.com.
\describe{ As as example: \code{Hostname = "hostname", Domainname = "domain.com"} would correspond to \code{docker run -h hostname.domain.com}}}

\item{User}{A string value containg the user to use inside the container. Corresponds to docker run -u}

\item{Memory}{Memory limit in bytes. Corresponds to docker run -m}

\item{MemorySwap}{Total memory usage (memory + swap); set -1 to disable swap.}

\item{CpuShares}{An integer value containing the CPU Shares for container (ie. the relative weight vs othercontainers). Corresponds to docker run -c}

\item{Cpuset}{String value containg the cgroups Cpuset to use. Corresponds to docker run --cpuset=""}

\item{AttachStdin}{Boolean value, attaches to stdin. Corresponds to docker run -a stdin}

\item{AttachStdout}{Boolean value, attaches to stdin. Corresponds to docker run -a stdout}

\item{AttachStderr}{Boolean value, attaches to stdin. Corresponds to docker run -a stderr}
}
\description{
\code{containerOpt}
A utility function to create options suitable for creating a container.
}
\examples{
\dontrun{
containerOpt()
}
}

