% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{docker-class}
\alias{docker}
\alias{docker-class}
\title{docker Class uses the}
\description{
docker is a generator object.
}
\section{Methods}{

\describe{
\item{\code{getContainers(all = TRUE, limit = NULL, since = NULL, before = NULL,
  size = NULL)}}{List containers:
\describe{
\item{\code{all}:}{1/True/true or 0/False/false, Show all containers. Only running containers are shown by default (i.e., this defaults to false)}
\item{\code{limit}:}{Show limit last created containers, include non-running ones.}
\item{\code{since}:}{Show only containers created since Id, include non-running ones.}
\item{\code{before}:}{Show only containers created before Id, include non-running ones.}
\item{\code{size}:}{1/True/true or 0/False/false, Show the containers sizes}
}}
}}
\examples{
\dontrun{
docker <- docker(ip = "192.168.59.103", port = 2375L) # windows example
docker$get()
}
}

