% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{dockerContainer-class}
\alias{dockerContainer}
\alias{dockerContainer-class}
\title{docker Class uses the}
\description{
dockerContainer is a generator object.
}
\section{Methods}{

\describe{
\item{\code{export(Class)}}{Returns the result of coercing the object to
Class.  No effect on the object itself.}

\item{\code{fsChanges()}}{Inspect changes on containers filesystem.}

\item{\code{inspect()}}{Return low-level information on the container.}

\item{\code{listProcesses(ps_args = NULL)}}{List processes running inside the container.
\describe{
\item{\code{ps_args}:}{ps arguments to use (e.g., aux). docker os dependent see \url{https://github.com/docker/docker/issues/8075}}
}}

\item{\code{logs(stdout = FALSE, stderr = FALSE, timestamps = FALSE, tail = "all")}}{Get stdout and stderr logs from the container.
\describe{
\item{\code{stdout}:}{1/True/true or 0/False/false, show stdout log. Default false}
\item{\code{stderr}:}{1/True/true or 0/False/false, show stderr log. Default false}
\item{\code{timestamps}:}{1/True/true or 0/False/false, print timestamps for every log line. Default false}
\item{\code{tail}:}{Output specified number of lines at the end of logs: all or <number>. Default all}
}}

\item{\code{resize(height = NULL, width = NULL)}}{Resize the TTY of the container.
\describe{
\item{\code{height}:}{Height for the resized container.}
\item{\code{width}:}{Width for the resized container.}
}}

\item{\code{restart(t = NULL)}}{Restart the container.
\describe{
 \item{\code{t}:}{number of seconds to wait before killing the container.}
}}

\item{\code{stop(t = NULL)}}{Stop the container.
\describe{
 \item{\code{t}:}{number of seconds to wait before killing the container.}
}}
}}
\examples{
\dontrun{
}
}

